+++ packages/react-native-alipay/android/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <!-- AndroidManifest.xml -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactlibrary">
+          package="com.thundersdata.alipay">
 
 </manifest>
