+++ packages/react-native-alipay/android/src/main/java/com/thundersdata/alipay/ReactNativeAlipayPackage.java
@@ -1,6 +1,6 @@
 // ReactNativeAlipayPackage.java
 
-package com.reactlibrary;
+package com.thundersdata.alipay;
 
 import java.util.Arrays;
 import java.util.Collections;
