+++ packages/react-native-image-picker/example/android/.settings/org.eclipse.buildship.core.prefs
@@ -5,9 +5,9 @@ connection.gradle.distribution=GRADLE_DISTRIBUTION(WRAPPER)
 connection.project.dir=
 eclipse.preferences.version=1
 gradle.user.home=
-java.home=
+java.home=/Library/Java/JavaVirtualMachines/jdk1.8.0_261.jdk/Contents/Home
 jvm.arguments=
 offline.mode=false
-override.workspace.settings=false
-show.console.view=false
-show.executions.view=false
+override.workspace.settings=true
+show.console.view=true
+show.executions.view=true
